[Inductor][NVGEMM] Enable nvMatmulHeuristics for FP4 blockscaled GEMM
Map torch.float4_e2m1fn_x2 to nvMatmulHeuristics type abbreviation
"F4" (multi-character token). FP4 precision strings use the 3-letter
format (e.g., F4F4S for FP4*FP4->FP32) since "F4" cannot be used in
the 5-letter single-char format.

Authored with Claude.

[ghstack-poisoned]
